feat(keyboard): sofle - make tap-hold more reliable
Add permissive hold and set tapping term to a crazy long value to make
accidental holds very difficult.  I want tap/hold determination to be
mostly independent of my typing speed.  Instead I want to rely on
ensuring that I always keep the tap/hold key depressed until the target
key is released, to ensure the permissive hold rule is triggered.  This
keeps holds snappy as they can trigger as soon as the target key is
released as opposed to waiting for the tapping term to expire.